+++ django_functest/__init__.py
@@ -1,3 +1,4 @@
+from .base import FuncBaseMixin
 from .funcselenium import FuncSeleniumMixin
 from .funcwebtest import FuncWebTestMixin
 from .utils import AdminLoginMixin, ShortcutLoginMixin
@@ -5,4 +6,4 @@ from .files import Upload
 
 __version__ = '0.1.9-dev'
 
-__all__ = ['FuncWebTestMixin', 'FuncSeleniumMixin', 'ShortcutLoginMixin', 'AdminLoginMixin', 'Upload']
+__all__ = ['FuncBaseMixin', 'FuncWebTestMixin', 'FuncSeleniumMixin', 'ShortcutLoginMixin', 'AdminLoginMixin', 'Upload']