+++ django_functest/tests/test_utils.py
@@ -1,6 +1,9 @@
+import inspect
+from unittest import TestCase
+
 from django.contrib.auth import get_user_model
 
-from django_functest import AdminLoginMixin, ShortcutLoginMixin
+from django_functest import AdminLoginMixin, FuncBaseMixin, FuncSeleniumMixin, FuncWebTestMixin, ShortcutLoginMixin
 
 from .base import ChromeBase, FirefoxBase, PhantomJSBase, WebTestBase
 
@@ -91,3 +94,29 @@ class TestAdminLoginChrome(TestAdminLoginBase, ChromeBase):
 
 class TestAdminLoginPhantomJS(TestAdminLoginBase, PhantomJSBase):
     pass
+
+
+class TestDocStrings(TestCase):
+    def test_doc_strings(self):
+        bad_docstrings = []
+        for cls in [FuncSeleniumMixin, FuncWebTestMixin]:
+            for name, member in inspect.getmembers(cls):
+                if name.startswith('__'):
+                    continue
+                if hasattr(member, 'im_func'):
+                    member = member.im_func  # Python 2 method
+                member_doc = getattr(member, '__doc__', None)
+
+                base_member = getattr(FuncBaseMixin, name, None)
+                if base_member is not None:
+                    if hasattr(base_member, 'im_func'):
+                        base_member = base_member.im_func  # Python 2 method
+                    base_doc = getattr(base_member, '__doc__', None)
+                    if base_doc is not None and member_doc != base_doc:
+                        bad_docstrings.append((cls, name))
+
+        if bad_docstrings:
+            self.fail("The following methods have incorrect or missing docstrings "
+                      "compared to FuncBaseMixin: \n" +
+                      "\n".join("{0}.{1}".format(cls.__name__, name)
+                                for cls, name in bad_docstrings))